refactor(wpclient): decouple UI rendering, event handling, and state management in tui.rs (close #44)
